+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# unreleased 
+- Add integration commands
+
 # v0.21.5 
 - Fix a bug where stream responses were not correctly parsed 
 - Fix a bug where streams were not correctly advanced